- Unbreak with latest version of go - Add deprecation notice - Bump PORTREVISION

+++ sysutils/logstash-forwarder/files/patch-fileinfo_freebsd.go
@@ -8,7 +8,7 @@
 +  "syscall"
 +)
 +
-+func file_ids(info *os.FileInfo) (uint32, uint32) {
++func file_ids(info *os.FileInfo) (uint64, uint64) {
 +  fstat := (*(info)).Sys().(*syscall.Stat_t)
 +  return fstat.Ino, fstat.Dev
 +}
