+++ sysutils/logstash-forwarder/files/patch-filestate_freebsd.go
@@ -6,6 +6,6 @@
 +type FileState struct {
 +  Source *string `json:"source,omitempty"`
 +  Offset int64 `json:"offset,omitempty"`
-+  Inode uint32 `json:"inode,omitempty"`
-+  Device uint32 `json:"device,omitempty"`
++  Inode uint64 `json:"inode,omitempty"`
++  Device uint64 `json:"device,omitempty"`
 +}
